Extract light device session modal
