+++ js/light-devices.js
@@ -24,6 +24,7 @@ import { recordTombstone } from './data-merge.js';
 import { CHANNEL_DISPLAY, BODY_REGIONS } from './sun.js';
 import { callClaudeAPI, hasAIProvider, supportsVision } from './api.js';
 import { resizeImage, isValidImageType, formatImageBlock, buildVisionContent } from './image-utils.js';
+import { openDeviceSessionDialog as openDeviceSessionDialogModal } from './light-device-session-modal.js';
 
 // Preset library is loaded lazily — keeps the JSON out of the boot path.
 let _PRESETS = null;
@@ -1449,302 +1450,13 @@ export async function addCustomDevice(spec) {
 // ─── UI: log device session modal ──────────────────────────────────────
 
 export async function openDeviceSessionDialog(deviceId) {
-  // Lazy hydrate — covers the case where the boot-time migration was
-  // skipped (page opened mid-init, presets cache cold, etc.) so the
-  // dialog always renders with the latest preset schema. Idempotent;
-  // no-op once devices carry channelGroups / modes / coupling.
-  await hydrateDevicesFromPresets().catch(() => {});
-  const device = getDevices().find(d => d.id === deviceId);
-  if (!device) return;
-  // Prefill from the user's last logged session on this device. First-
-  // time logs fall through to vendor reference distance + sensible
-  // defaults (10 min, torso, eyes protected). Each save updates
-  // device.lastSession so the dialog opens with the user's actual
-  // ritual next time.
-  const last = device.lastSession || {};
-  const defaultDuration = Number.isFinite(last.durationMin) && last.durationMin > 0 ? last.durationMin : 10;
-  const defaultDistanceCm = Number.isFinite(last.distanceCm) && last.distanceCm > 0
-    ? last.distanceCm
-    : (device.recommendedDistanceCm || 15);
-  const defaultEyesProtected = last.eyesProtected !== false;
-  // bodyAreas[] is the new precise per-region field. For legacy
-  // sessions that only have a broad bodyArea string, expand it to the
-  // matching region keys so the silhouette pre-selects sensibly.
-  const BROAD_TO_REGIONS = {
-    face:  ['face'],
-    torso: ['breast-chest', 'torso-front', 'abdomen'],
-    arms:  ['arms-front', 'arms-back'],
-    legs:  ['legs-front', 'legs-back'],
-    // Legacy keys preserved for backcompat reads (last.bodyArea may
-    // still be 'whole-body' or 'targeted' from pre-toggle sessions).
-    'whole-body': (BODY_REGIONS || []).map(r => r.key),
-    targeted: ['breast-chest'],
-  };
-  let defaultRegions;
-  if (Array.isArray(last.bodyAreas) && last.bodyAreas.length > 0) {
-    defaultRegions = last.bodyAreas.slice();
-  } else if (last.bodyArea && BROAD_TO_REGIONS[last.bodyArea]) {
-    defaultRegions = BROAD_TO_REGIONS[last.bodyArea].slice();
-  } else {
-    defaultRegions = ['breast-chest'];  // sensible single-region default
-  }
-  // Mode picker — only renders for devices with multiple modes (Maxi UVB,
-  // Trinity, etc.). Coupling-violating modes are filtered out so users
-  // can't pick the device into an unsafe state from the in-modal picker
-  // (e.g. Maxi UVB has no UV-only entry; D-Light on Trinity stays
-  // available since Trinity has no coupling rules).
-  const validateMode = window.validateModeCoupling || (() => ({ ok: true }));
-  const validModes = Array.isArray(device.modes)
-    ? device.modes.filter(m => validateMode(device, m.id).ok)
-    : [];
-  const showModePicker = validModes.length > 1;
-  let defaultMode = null;
-  if (showModePicker) {
-    const lastModeValid = last.mode && validModes.some(m => m.id === last.mode);
-    defaultMode = lastModeValid ? last.mode : (validModes.find(m => m.default) || validModes[0])?.id || null;
-  }
-  const overlay = document.createElement('div');
-  overlay.className = 'modal-overlay show';
-  overlay.innerHTML = `<div class="modal" role="dialog" aria-label="Log device session">
-    <div class="modal-header">
-      <h3>Log session — ${escapeHTML(device.brand)} ${escapeHTML(device.model)}</h3>
-      <button class="modal-close" onclick="this.closest('.modal-overlay').remove()" aria-label="Close">×</button>
-    </div>
-    <div class="modal-body">
-      ${showModePicker ? `
-        <div class="ctx-label dev-mode-field">
-          <span>Mode</span>
-          <input type="hidden" id="dev-session-mode" value="${escapeAttr(defaultMode || '')}" />
-          <div class="dev-mode-picker" role="radiogroup" aria-label="Device mode">
-            ${validModes.map(m => `<button type="button" class="dev-mode-btn${m.id === defaultMode ? ' active' : ''}" data-mode="${escapeAttr(m.id)}" role="radio" aria-checked="${m.id === defaultMode ? 'true' : 'false'}" title="Which LED groups were firing for this session — picked from the device's vendor-defined modes. Affects channel-dose math.">${escapeHTML(m.label || m.id)}</button>`).join('')}
-          </div>
-        </div>
-      ` : ''}
-      <label class="ctx-label">Duration (minutes)
-        <input type="number" id="dev-session-duration" class="ctx-input" min="1" max="120" value="${defaultDuration}" />
-      </label>
-      ${(() => {
-        const useUS = state.unitSystem === 'US';
-        const startUnit = useUS ? 'in' : 'cm';
-        const refCm = device.recommendedDistanceCm || 15;
-        const fmt = (cm, u) => u === 'in' ? +(cm / 2.54).toFixed(1) : cm;
-        // Hint surfaces the vendor reference always; if the user has
-        // overridden it before, surface that override too so they know
-        // why the input shows a different default.
-        const hasOverride = Number.isFinite(last.distanceCm) && Math.abs(last.distanceCm - refCm) > 0.5;
-        const overrideHint = hasOverride
-          ? ` You usually log at ${fmt(defaultDistanceCm, 'cm')} cm — prefilled below.`
-          : '';
-        return `<label class="ctx-label">Distance from device
-          <div class="dev-distance-row">
-            <input type="number" id="dev-session-distance" class="ctx-input" min="2" max="200" step="0.5" value="${fmt(defaultDistanceCm, startUnit)}" data-unit="${startUnit}" data-base-cm="${refCm}" />
-            <div class="dev-unit-toggle" role="tablist" aria-label="Distance unit">
-              <button type="button" class="dev-unit-btn${startUnit === 'cm' ? ' active' : ''}" data-unit="cm" role="tab" aria-selected="${startUnit === 'cm'}">cm</button>
-              <button type="button" class="dev-unit-btn${startUnit === 'in' ? ' active' : ''}" data-unit="in" role="tab" aria-selected="${startUnit === 'in'}">in</button>
-            </div>
-          </div>
-          <span class="dev-session-hint">Vendor reference: ${fmt(refCm, 'cm')} cm (${fmt(refCm, 'in')} in).${overrideHint} The dose math uses inverse-square scaling around this point — close ranges magnify errors fast.</span>
-        </label>`;
-      })()}
-      <div class="ctx-label" style="display:block">
-        <span>Body area treated</span>
-        <div class="sun-silhouette-wrap" id="dev-session-silhouette-slot">${(typeof window !== 'undefined' && window.renderBodySilhouette) ? window.renderBodySilhouette(new Set(defaultRegions)) : ''}</div>
-        <div class="sun-silhouette-hint-row" style="display:flex;align-items:center;justify-content:space-between;gap:8px">
-          <div class="sun-silhouette-hint" id="dev-session-area-hint">Tap regions the panel reaches.</div>
-          <button type="button" class="ctx-btn-option" id="dev-session-clear" style="padding:2px 10px;font-size:11px">Clear</button>
-        </div>
-      </div>
-      <div class="ctx-label" style="display:flex;align-items:center;justify-content:space-between;gap:12px">
-        <span style="flex:1;min-width:0">Eyes protected (goggles or closed)</span>
-        <label class="toggle-switch">
-          <input type="checkbox" id="dev-session-eyes"${defaultEyesProtected ? ' checked' : ''} />
-          <span class="toggle-slider"></span>
-        </label>
-      </div>
-      <p class="modal-body-hint" style="margin-top:8px">Save now to log a finished session, or Start to run a live timer (matches the sun-session pattern — handy when you want to walk away and come back).</p>
-      <div class="modal-actions" style="margin-top:18px">
-        <button class="import-btn import-btn-secondary" onclick="this.closest('.modal-overlay').remove()">Cancel</button>
-        <button class="import-btn import-btn-secondary" id="dev-session-start">Start timer</button>
-        <button class="import-btn import-btn-primary" id="dev-session-save">Save session</button>
-      </div>
-    </div>
-  </div>`;
-  _wireModal(overlay);
-
-  let lastModePointerActivation = 0;
-  const setMode = (btn) => {
-    const mode = btn.dataset.mode || '';
-    const input = overlay.querySelector('#dev-session-mode');
-    if (input) input.value = mode;
-    for (const b of overlay.querySelectorAll('.dev-mode-btn[data-mode]')) {
-      const active = b === btn;
-      b.classList.toggle('active', active);
-      b.setAttribute('aria-checked', active ? 'true' : 'false');
-    }
-  };
-  for (const btn of overlay.querySelectorAll('.dev-mode-btn[data-mode]')) {
-    btn.addEventListener('pointerup', (e) => {
-      if (e.pointerType === 'mouse') return;
-      setMode(btn);
-      lastModePointerActivation = Date.now();
-      e.preventDefault();
-    }, { passive: false });
-    btn.addEventListener('touchend', (e) => {
-      if (Date.now() - lastModePointerActivation < 80) return;
-      setMode(btn);
-      lastModePointerActivation = Date.now();
-      e.preventDefault();
-    }, { passive: false });
-    btn.addEventListener('click', (e) => {
-      if (Date.now() - lastModePointerActivation < 700) {
-        e.preventDefault();
-        return;
-      }
-      setMode(btn);
-    });
-  }
-
-  // Silhouette wiring — pre-selected regions plus an updateHint
-  // callback that recomputes the % skin coverage shown under the
-  // figure. Pre-2026-05-08 this was a 6-option dropdown ('torso',
-  // 'whole-body', etc.); now it's the same 16-region picker as sun
-  // sessions for consistency with the user's mental model.
-  const selectedRegions = new Set(defaultRegions);
-  const _fracByKey = Object.fromEntries((BODY_REGIONS || []).map(r => [r.key, r.fraction]));
-  const _labelByKey = Object.fromEntries((BODY_REGIONS || []).map(r => [r.key, r.label]));
-  const _silhouetteSlot = overlay.querySelector('#dev-session-silhouette-slot');
-  const _hint = overlay.querySelector('#dev-session-area-hint');
-  function _updateAreaHint(set) {
-    if (!_hint) return;
-    if (set.size === 0) {
-      _hint.textContent = 'Pick at least one region — what does the panel reach?';
-      return;
-    }
-    const frac = Array.from(set).reduce((s, k) => s + (_fracByKey[k] || 0), 0);
-    const labels = Array.from(set).map(k => _labelByKey[k] || k).slice(0, 4).join(', ');
-    const more = set.size > 4 ? ` +${set.size - 4} more` : '';
-    _hint.textContent = `${set.size} region${set.size === 1 ? '' : 's'} (~${Math.round(frac * 100)}% of skin) — ${labels}${more}`;
-  }
-  if (_silhouetteSlot && typeof window !== 'undefined' && window.bindBodySilhouette) {
-    window.bindBodySilhouette(_silhouetteSlot, selectedRegions, (set) => {
-      _updateAreaHint(set);
-    });
-  }
-  _updateAreaHint(selectedRegions);
-
-  // Clear — single bulk-deselect affordance. The silhouette itself is
-  // the picker; pre-2026-05-08 also had a 4-zone toggle row but it was
-  // a redundant abstraction (every operation duplicated a silhouette
-  // tap, and active-state drift on individual taps was confusing).
-  // Real logs show 1-3 region picks, not whole-zone bulk operations,
-  // so this single Clear button covers the only genuine shortcut.
-  overlay.querySelector('#dev-session-clear')?.addEventListener('click', () => {
-    selectedRegions.clear();
-    if (_silhouetteSlot && typeof window !== 'undefined' && window.renderBodySilhouette) {
-      _silhouetteSlot.innerHTML = window.renderBodySilhouette(selectedRegions);
-    }
-    _updateAreaHint(selectedRegions);
-  });
-
-  // Per-field unit toggle: cm ↔ in. Lets a US user briefly type a cm
-  // value (or vice versa) without mental math when their global unit
-  // preference doesn't match the spec sheet they're reading from.
-  // Conversion happens in-place on the visible value; data-unit attr
-  // tracks what the field is currently representing.
-  for (const btn of overlay.querySelectorAll('.dev-unit-btn')) {
-    btn.addEventListener('click', () => {
-      const target = btn.getAttribute('data-unit');
-      const input = overlay.querySelector('#dev-session-distance');
-      const cur = input.dataset.unit || 'cm';
-      if (cur === target) return;
-      const v = parseFloat(input.value);
-      if (Number.isFinite(v)) {
-        const cm = cur === 'in' ? v * 2.54 : v;
-        input.value = target === 'in' ? +(cm / 2.54).toFixed(1) : Math.round(cm);
-      }
-      input.dataset.unit = target;
-      input.step = target === 'in' ? '0.5' : '1';
-      for (const b of overlay.querySelectorAll('.dev-unit-btn')) {
-        const active = b.getAttribute('data-unit') === target;
-        b.classList.toggle('active', active);
-        b.setAttribute('aria-selected', active ? 'true' : 'false');
-      }
-    });
-  }
-
-  overlay.querySelector('#dev-session-save').addEventListener('click', async () => {
-    const durationMin = parseInt(overlay.querySelector('#dev-session-duration').value, 10) || 10;
-    // Read distance in whatever unit the input was rendered with — the
-    // data-unit attribute carries the unit, the dose math always works in cm.
-    const distInput = overlay.querySelector('#dev-session-distance');
-    const distVal = parseFloat(distInput.value);
-    const distUnit = distInput.dataset.unit || 'cm';
-    const distanceCm = Number.isFinite(distVal)
-      ? (distUnit === 'in' ? distVal * 2.54 : distVal)
-      : (device.recommendedDistanceCm || 15);
-    const bodyAreas = Array.from(selectedRegions);
-    if (bodyAreas.length === 0) {
-      _updateAreaHint(selectedRegions);
-      _hint?.classList.add('sun-silhouette-hint-error');
-      setTimeout(() => _hint?.classList.remove('sun-silhouette-hint-error'), 2500);
-      return;
-    }
-    // Denormalized broad-zone hint kept for legacy listing rows that
-    // haven't been migrated to bodyAreas yet. Pick the simplest match
-    // for the chosen region set.
-    let bodyArea = 'targeted';
-    if (bodyAreas.length >= (BODY_REGIONS || []).length - 2) bodyArea = 'whole-body';
-    else if (bodyAreas.every(r => r.startsWith('legs') || r.startsWith('feet'))) bodyArea = 'legs';
-    else if (bodyAreas.every(r => r.startsWith('arms'))) bodyArea = 'arms';
-    else if (bodyAreas.every(r => /face|thyroid/.test(r))) bodyArea = 'face';
-    else if (bodyAreas.every(r => /chest|torso|abdomen|breast/.test(r))) bodyArea = 'torso';
-    const eyesProtected = overlay.querySelector('#dev-session-eyes').checked;
-    const mode = showModePicker ? overlay.querySelector('#dev-session-mode')?.value || null : null;
-    await logDeviceSession({ deviceId, durationMin, distanceCm, bodyArea, bodyAreas, eyesProtected, mode });
-    overlay.remove();
-    showNotification(`${durationMin} min ${escapeHTML(device.brand)} session saved.`);
-    // Always land on /light after a save so the user sees the freshly-
-    // logged session (and its mode chip) in their history. Pre-Round-7
-    // this only re-rendered when already on /light; saving from the
-    // dashboard FAB used to leave the user on the dashboard with no
-    // visual feedback that the session had been recorded.
-    if (window.navigate) window.navigate('light');
-  });
-
-  // Start-timer path — same shared form, but begins a live session
-  // instead of logging a finished one. Reuses every input EXCEPT
-  // duration (irrelevant — duration accumulates from startedAt).
-  overlay.querySelector('#dev-session-start').addEventListener('click', async () => {
-    if (getActiveDeviceSession()) {
-      showNotification('Another device session is already running. Stop it first.', 'error');
-      return;
-    }
-    const distInput = overlay.querySelector('#dev-session-distance');
-    const distVal = parseFloat(distInput.value);
-    const distUnit = distInput.dataset.unit || 'cm';
-    const distanceCm = Number.isFinite(distVal)
-      ? (distUnit === 'in' ? distVal * 2.54 : distVal)
-      : (device.recommendedDistanceCm || 15);
-    const bodyAreas = Array.from(selectedRegions);
-    if (bodyAreas.length === 0) {
-      _updateAreaHint(selectedRegions);
-      _hint?.classList.add('sun-silhouette-hint-error');
-      setTimeout(() => _hint?.classList.remove('sun-silhouette-hint-error'), 2500);
-      return;
-    }
-    let bodyArea = 'targeted';
-    if (bodyAreas.length >= (BODY_REGIONS || []).length - 2) bodyArea = 'whole-body';
-    else if (bodyAreas.every(r => r.startsWith('legs') || r.startsWith('feet'))) bodyArea = 'legs';
-    else if (bodyAreas.every(r => r.startsWith('arms'))) bodyArea = 'arms';
-    else if (bodyAreas.every(r => /face|thyroid/.test(r))) bodyArea = 'face';
-    else if (bodyAreas.every(r => /chest|torso|abdomen|breast/.test(r))) bodyArea = 'torso';
-    const eyesProtected = overlay.querySelector('#dev-session-eyes').checked;
-    const mode = showModePicker ? overlay.querySelector('#dev-session-mode')?.value || null : null;
-    await startDeviceSession({ deviceId, distanceCm, bodyAreas, bodyArea, eyesProtected, mode });
-    overlay.remove();
-    showNotification(`Live ${escapeHTML(device.brand)} session started — tap Stop & save when finished.`);
-    ensureActiveDeviceTicker();
-    if (window.navigate) window.navigate('light');
+  return openDeviceSessionDialogModal(deviceId, {
+    hydrateDevicesFromPresets,
+    getDevices,
+    logDeviceSession,
+    getActiveDeviceSession,
+    startDeviceSession,
+    ensureActiveDeviceTicker,
   });
 }
