+++ service-worker.js
@@ -285,6 +285,7 @@ const APP_SHELL = [
   '/js/light-channels-ai-analysis.js',
   '/js/light-device-ai-analysis.js',
   '/js/light-devices.js',
+  '/js/light-device-session-modal.js',
   '/js/light-env.js',
   '/js/light-env-ai-analysis.js',
   '/js/light-screen-ai-analysis.js',
