+++ tests/test-a11y-phase3.js
@@ -194,6 +194,9 @@ console.log('=== Phase 3 A11y Tests ===\n');
   // modals (Log device session) require explicit Cancel/Save so accidental
   // taps don't lose typed values.
   const lightDevSrc = read('/js/light-devices.js');
+  const lightDevSessionSrc = read('/js/light-device-session-modal.js');
+  assert('light-devices.js delegates session dialog rendering to extracted module',
+    lightDevSrc.includes("from './light-device-session-modal.js'"));
   // Two browse modals each get a backdrop-close listener guarded by
   // `e.target === overlay` so child clicks don't bubble out.
   const backdropMatches = lightDevSrc.match(/overlay\.addEventListener\('click', \(e\) => \{\s*if \(e\.target === overlay\) overlay\.remove\(\);/g) || [];
@@ -202,9 +205,9 @@ console.log('=== Phase 3 A11y Tests ===\n');
     `found ${backdropMatches.length} backdrop-close listeners`);
   // openDeviceSessionDialog is a form modal — must NOT have backdrop-close
   // (would lose typed duration/distance/notes on stray click).
-  const sessionDialogStart = lightDevSrc.indexOf('export async function openDeviceSessionDialog');
-  const sessionDialogEnd = lightDevSrc.indexOf('export', sessionDialogStart + 1);
-  const sessionDialogBody = lightDevSrc.slice(sessionDialogStart, sessionDialogEnd > 0 ? sessionDialogEnd : undefined);
+  const sessionDialogStart = lightDevSessionSrc.indexOf('export async function openDeviceSessionDialog');
+  const sessionDialogEnd = lightDevSessionSrc.indexOf('export', sessionDialogStart + 1);
+  const sessionDialogBody = lightDevSessionSrc.slice(sessionDialogStart, sessionDialogEnd > 0 ? sessionDialogEnd : undefined);
   assert('openDeviceSessionDialog (form modal) has NO backdrop-close listener',
     !/overlay\.addEventListener\('click'/.test(sessionDialogBody));
